Move `SvgIcon` component to `src/shared/components`
Move SvgIcon component to `src/shared/` in preparation for re-use in the
"annotator" part of the application.
