+++ src/shared/components/svg-icon.js
