+++ src/shared/components/test/svg-icon-test.js
