+++ src/sidebar/components/annotation-header.js
@@ -8,7 +8,7 @@ import AnnotationDocumentInfo from './annotation-document-info';
 import AnnotationShareInfo from './annotation-share-info';
 import AnnotationUser from './annotation-user';
 import Button from './button';
-import SvgIcon from './svg-icon';
+import SvgIcon from '../../shared/components/svg-icon';
 import Timestamp from './timestamp';
 
 /**
