+++ src/sidebar/components/sidebar-content-error.js
@@ -4,7 +4,7 @@ import propTypes from 'prop-types';
 import useStore from '../store/use-store';
 
 import Button from './button';
-import SvgIcon from './svg-icon';
+import SvgIcon from '../../shared/components/svg-icon';
 
 /**
  * An error message to display in the sidebar.
